+++ dlrover/python/elastic_agent/master_client.py
@@ -386,7 +386,6 @@ class MasterClient(Singleton):
 
     def report_diagnosis_cuda_log(self, cuda_log: CudaLog):
         message = grpc.DiagnosisCudaLog(cuda_log.timestamp)
-        print("report cuda log!!!")
         self._report(message)
 
     def get_paral_config(self) -> grpc.ParallelConfig:
